tools/PerformanceProfiler.pack/requirements.txt
google-cloud-bigquery==1.26.1
google-cloud-storage==1.30.0
moz-sql-parser==4.21.21059
